SampleProject/prateek_gupta/requirements.txt
Django~=5.0.10
pandas~=2.2.3
javaproperties~=0.8.2
PyJWT~=2.8.0
mysql~=0.0.3
asgiref~=3.8.1
boto3~=1.37.1
botocore~=1.37.1
aioboto3~=14.1.0
aiobotocore~=2.21.1
confluent-kafka~=2.9.0
uvicorn~=0.34.0
mysql-connector-python~=9.2.0
aiokafka~=0.12.0
redis~=5.2.1
certifi~=2024.12.14
typing_extensions~=4.12.2
urllib3~=2.3.0
idna~=3.10
charset-normalizer~=3.4.0
requests~=2.32.3
asyncio~=3.4.3
opensearch-py~=2.8.0